refactor: changed resize_image to receive a scale in float

+++ sources_bonus/graphics/graphics_bonus.h
@@ -6,7 +6,7 @@
 /*   By: wleite <wleite@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/05 05:46:11 by jofelipe          #+#    #+#             */
-/*   Updated: 2022/02/20 00:53:01 by wleite           ###   ########.fr       */
+/*   Updated: 2022/02/23 17:55:43 by wleite           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,6 +46,6 @@ void	copy_layer(t_img *from, t_img *to, int x, int y);
 int		blend_colors(int color1, int color2, float alpha1, float alpha2);
 void	change_color_intensity(int *color, float factor);
 int		get_color(int r, int g, int b);
-t_img	*resize_image(t_img *image_to_scale, int scale, t_data *data);
+t_img	*resize_image(t_img *image_to_scale, float scale, t_data *data);
 
 #endif
